+++ readwisereader.koplugin/constants.lua
@@ -1,0 +1,7 @@
+-- SPDX-License-Identifier: AGPL-3.0-only
+
+return {
+    VERSION = "0.0.2",
+    SETTINGS_FILENAME = "readwisereader.lua",
+    AUTH_URL = "https://readwise.io/api/v2/auth/",
+}